requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
tensorflow
python-dotenv
requests
scikit-learn
exif
opencv-python-headless
supabase
matplotlib
transformers
torch
torchvision